requirements.txt
Django==1.6
PyJWT==1.3.0
South==1.0.2
Unipath==1.1
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-braces==1.8.0
django-toolbelt==0.0.1
gunicorn==19.3.0
oauthlib==0.7.2
psycopg2==2.6
python-openid==2.2.5
python-social-auth==0.2.10
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
static3==0.6.1
wsgiref==0.1.2
